Call waitIdle after each frame
This is a temporary workaround for a synchronisation issue in the
renderer. This issue shows under Linux where the application hangs for
multiple seconds when changing focus or moving windows around.

Since we are already GPU bound this only has marginal impact on the
framerate. We plan to resolve this workaround for version 1.2.

ref #2